requirements.txt
numpy>=1.23
scipy>=1.10
pandas>=2.0
matplotlib>=3.7
pyyaml>=6.0
scikit-learn>=1.3

# ML stack (uncomment sesuai kebutuhan)
# torch>=2.0
# pytorch-lightning>=2.0
